word-format-tool/requirements.txt
strestreamlit>=1.20.0
python-docx>=0.8.11
pandas>=1.5.0
requests>=2.28.0
textract>=1.6.5
pdfplumber>=0.10.0
